DeleteOrganizationTest
